Move prettyprint out of action_lib
It isn't used by action_lib, and codewise is fairly different than action_lib
