+++ ext/dcos-installer/dcos_installer/prettyprint.py
